feat(uniform exploreres): Making explorers uniform
