+++ pan3d/ui/preview.py
@@ -1,261 +1,19 @@
 import math
 from pathlib import Path
 
-from trame.decorators import TrameApp, change
+from trame.decorators import change
 from trame.widgets import html, vuetify3 as v3
 
 from pan3d import catalogs as pan3d_catalogs
 from pan3d.utils.constants import XYZ, SLICE_VARS
 from pan3d.utils.convert import max_str_length
-from pan3d.utils.presets import PRESETS
 
-from pan3d.ui.css import base, preview
-from pan3d.ui.collapsible import CollapsableSection
+from pan3d.utils.common import RenderingSettingsBasic
 
 
-class SummaryToolbar(v3.VCard):
-    def __init__(
-        self,
-        t_labels="t_labels",
-        slice_t="slice_t",
-        slice_t_max="slice_t_max",
-        color_by="color_by",
-        data_arrays="data_arrays",
-        max_time_width="max_time_width",
-        max_time_index_width="max_time_index_width",
-        **kwargs,
-    ):
-        super().__init__(
-            classes="summary-toolbar",
-            rounded="pill",
-            **kwargs,
-        )
-
-        # Activate CSS
-        self.server.enable_module(base)
-        self.server.enable_module(preview)
-
-        with self:
-            with v3.VToolbar(
-                classes="pl-2",
-                height=50,
-                elevation=1,
-                style="background: none;",
-            ):
-                v3.VIcon("mdi-clock-outline")
-                html.Pre(
-                    f"{{{{ {t_labels}[slice_t] }}}}",
-                    classes="mx-2 text-left",
-                    style=(f"`min-width: ${{ {max_time_width} }}rem;`",),
-                )
-                v3.VSlider(
-                    prepend_inner_icon="mdi-clock-outline",
-                    v_model=(slice_t, 0),
-                    min=0,
-                    max=(slice_t_max, 0),
-                    step=1,
-                    hide_details=True,
-                    density="compact",
-                    flat=True,
-                    variant="solo",
-                    classes="mx-2",
-                )
-                html.Div(
-                    f"{{{{ {slice_t} + 1 }}}}/{{{{ {slice_t_max} + 1 }}}}",
-                    classes="mx-2 text-right",
-                    style=(f"`min-width: ${{ {max_time_index_width} }}rem;`",),
-                )
-                v3.VSelect(
-                    placeholder="Color By",
-                    prepend_inner_icon="mdi-format-color-fill",
-                    v_model=(color_by, None),
-                    items=(data_arrays, []),
-                    clearable=True,
-                    hide_details=True,
-                    density="compact",
-                    flat=True,
-                    variant="solo",
-                    max_width=200,
-                )
-
-
-class DataOrigin(CollapsableSection):
-    def __init__(self, load_dataset):
-        super().__init__("Data origin", "show_data_origin", True)
-
-        self.state.load_button_text = "Load"
-        self.state.can_load = True
-        self.state.data_origin_id_to_desc = {}
-
-        with self.content:
-            v3.VSelect(
-                label="Source",
-                v_model=("data_origin_source", "xarray"),
-                items=(
-                    "data_origin_sources",
-                    pan3d_catalogs.list_availables(),
-                ),
-                hide_details=True,
-                density="compact",
-                flat=True,
-                variant="solo",
-            )
-            v3.VDivider()
-            v3.VTextField(
-                placeholder="Location",
-                v_if="['file', 'url'].includes(data_origin_source)",
-                v_model=("data_origin_id", ""),
-                hide_details=True,
-                density="compact",
-                flat=True,
-                variant="solo",
-                append_inner_icon=(
-                    "data_origin_id_error ? 'mdi-file-document-alert-outline' : undefined",
-                ),
-                error=("data_origin_id_error", False),
-            )
-
-            with v3.VTooltip(
-                v_else=True,
-                text=("`${ data_origin_id_to_desc[data_origin_id] }`",),
-            ):
-                with html.Template(v_slot_activator="{ props }"):
-                    v3.VSelect(
-                        v_bind="props",
-                        label="Name",
-                        v_model="data_origin_id",
-                        items=("data_origin_ids", []),
-                        hide_details=True,
-                        density="compact",
-                        flat=True,
-                        variant="solo",
-                    )
-
-            # v3.VDivider()
-            # v3.VSwitch(
-            #     label=("`Order ${data_origin_order}`",),
-            #     v_model=("data_origin_order", "C"),
-            #     true_value="C",
-            #     false_value="F",
-            #     hide_details=True,
-            #     density="compact",
-            #     flat=True,
-            #     variant="solo",
-            #     classes="mx-6",
-            # )
-            v3.VDivider()
-            v3.VBtn(
-                "{{ load_button_text }}",
-                block=True,
-                classes="text-none",
-                flat=True,
-                density="compact",
-                rounded=0,
-                disabled=("!data_origin_id?.length || !can_load",),
-                color=("can_load ? 'primary': undefined",),
-                click=(
-                    load_dataset,
-                    "[data_origin_source, data_origin_id, data_origin_order]",
-                ),
-            )
-
-
-class DataInformation(CollapsableSection):
-    def __init__(self, xarray_info="xarray_info"):
-        super().__init__("Data information", "show_data_information")
-
-        self._var_name = xarray_info
-        self.state.setdefault(xarray_info, [])
-
-        with self.content:
-            with v3.VTable(density="compact", hover=True):
-                with html.Tbody():
-                    with html.Template(v_for=f"item, i in {xarray_info}", key="i"):
-                        with v3.VTooltip():
-                            with html.Template(v_slot_activator="{ props }"):
-                                with html.Tr(v_bind="props", classes="pointer"):
-                                    with html.Td(
-                                        classes="d-flex align-center text-no-wrap"
-                                    ):
-                                        v3.VIcon(
-                                            "{{ item.icon }}",
-                                            size="sm",
-                                            classes="mr-2",
-                                        )
-                                        html.Div("{{ item.name }}")
-                                    html.Td(
-                                        "{{ item.length }}",
-                                        classes="text-right",
-                                    )
-
-                            with v3.VTable(
-                                density="compact",
-                                theme="dark",
-                                classes="no-bg ma-0 pa-0",
-                            ):
-                                with html.Tbody():
-                                    with html.Tr(
-                                        v_for="attr, j in item.attrs",
-                                        key="j",
-                                    ):
-                                        html.Td(
-                                            "{{ attr.key }}",
-                                        )
-                                        html.Td(
-                                            "{{ attr.value }}",
-                                        )
-
-    def update_information(self, xr, available_arrays=None):
-        xarray_info = []
-        coords = set(xr.coords.keys())
-        data = set(available_arrays or [])
-        for name in xr.variables:
-            icon = "mdi-variable"
-            order = 3
-            length = f'({",".join(xr[name].dims)})'
-            attrs = []
-            if name in coords:
-                icon = "mdi-ruler"
-                order = 1
-                length = xr[name].size
-                shape = xr[name].shape
-                if length > 1 and len(shape) == 1:
-                    attrs.append(
-                        {
-                            "key": "range",
-                            "value": f"[{xr[name].values[0]}, {xr[name].values[-1]}]",
-                        }
-                    )
-            if name in data:
-                icon = "mdi-database"
-                order = 2
-            xarray_info.append(
-                {
-                    "order": order,
-                    "icon": icon,
-                    "name": name,
-                    "length": length,
-                    "type": str(xr[name].dtype),
-                    "attrs": attrs
-                    + [
-                        {"key": "type", "value": str(xr[name].dtype)},
-                    ]
-                    + [
-                        {"key": str(k), "value": str(v)}
-                        for k, v in xr[name].attrs.items()
-                    ],
-                }
-            )
-        xarray_info.sort(key=lambda item: item["order"])
-
-        # Update UI
-        self.state[self._var_name] = xarray_info
-
-
-@TrameApp()
-class RenderingSettings(CollapsableSection):
-    def __init__(self, source, update_rendering):
-        super().__init__("Rendering", "show_rendering")
+class RenderingSettings(RenderingSettingsBasic):
+    def __init__(self, source, update_rendering, **kwargs):
+        super().__init__(source, update_rendering, **kwargs)
 
         self.source = source
         self.state.setdefault("slice_extents", {})
@@ -266,122 +24,6 @@ class RenderingSettings(CollapsableSection):
         self.state.setdefault("dataset_bounds", [0, 1, 0, 1, 0, 1])
 
         with self.content:
-            v3.VSelect(
-                placeholder="Data arrays",
-                prepend_inner_icon="mdi-database",
-                hide_selected=True,
-                v_model=("data_arrays", []),
-                items=("data_arrays_available", []),
-                multiple=True,
-                hide_details=True,
-                density="compact",
-                chips=True,
-                closable_chips=True,
-                flat=True,
-                variant="solo",
-            )
-            v3.VDivider()
-            v3.VSelect(
-                placeholder="Color By",
-                prepend_inner_icon="mdi-format-color-fill",
-                v_model=("color_by", None),
-                items=("data_arrays", []),
-                clearable=True,
-                hide_details=True,
-                density="compact",
-                flat=True,
-                variant="solo",
-            )
-            v3.VDivider()
-            with v3.VRow(no_gutters=True, classes="align-center mr-0"):
-                with v3.VCol():
-                    v3.VTextField(
-                        prepend_inner_icon="mdi-water-minus",
-                        v_model_number=("color_min", 0.45),
-                        type="number",
-                        hide_details=True,
-                        density="compact",
-                        flat=True,
-                        variant="solo",
-                        reverse=True,
-                    )
-                with v3.VCol():
-                    v3.VTextField(
-                        prepend_inner_icon="mdi-water-plus",
-                        v_model_number=("color_max", 5.45),
-                        type="number",
-                        hide_details=True,
-                        density="compact",
-                        flat=True,
-                        variant="solo",
-                        reverse=True,
-                    )
-                with html.Div(classes="flex-0"):
-                    v3.VBtn(
-                        icon="mdi-arrow-split-vertical",
-                        size="sm",
-                        density="compact",
-                        flat=True,
-                        variant="outlined",
-                        classes="mx-2",
-                        click=self.reset_color_range,
-                    )
-            # v3.VDivider()
-            with html.Div(classes="mx-2"):
-                html.Img(
-                    src=("preset_img", None),
-                    style="height: 0.75rem; width: 100%;",
-                    classes="rounded-lg border-thin",
-                )
-            v3.VSelect(
-                placeholder="Color Preset",
-                prepend_inner_icon="mdi-palette",
-                v_model=("color_preset", "Fast"),
-                items=("color_presets", list(PRESETS.keys())),
-                hide_details=True,
-                density="compact",
-                flat=True,
-                variant="solo",
-            )
-
-            with v3.VTooltip(
-                text=("`NaN Color (${nan_colors[nan_color]})`",),
-            ):
-                with html.Template(v_slot_activator="{ props }"):
-                    with v3.VItemGroup(
-                        v_model="nan_color",
-                        v_bind="props",
-                        classes="d-inline-flex ga-1 pa-2",
-                        mandatory="force",
-                    ):
-                        v3.VIcon(
-                            "mdi-eyedropper-variant",
-                            classes="my-auto mx-1 text-medium-emphasis",
-                        )
-                        with v3.VItem(
-                            v_for="(color, i) in nan_colors", key="i", value=("i",)
-                        ):
-                            with v3.Template(
-                                raw_attrs=['#default="{ isSelected, toggle }"']
-                            ):
-                                with v3.VAvatar(
-                                    density="compact",
-                                    color=("isSelected ? 'primary': 'transparent'",),
-                                ):
-                                    v3.VBtn(
-                                        "{{ color[3] < 0.1 ? 't' : '' }}",
-                                        density="compact",
-                                        border="md surface opacity-100",
-                                        color=(
-                                            "color[3] ? `rgb(${color[0] * 255}, ${color[1] * 255}, ${color[2] * 255})` : undefined",
-                                        ),
-                                        flat=True,
-                                        icon=True,
-                                        ripple=False,
-                                        size="small",
-                                        click="toggle",
-                                    )
-
             v3.VDivider()
             # X crop/cut
             with v3.VTooltip(
@@ -801,142 +443,3 @@ class RenderingSettings(CollapsableSection):
             self.state.color_by = None
 
         self.source.arrays = data_arrays
-
-
-class ControlPanel(v3.VCard):
-    def __init__(
-        self,
-        enable_data_selection,
-        source,
-        toggle,
-        load_dataset,
-        update_rendering,
-        import_file_upload,
-        export_file_download,
-        xr_update_info="xr_update_info",
-        source_update_rendering="source_update_rendering",
-        **kwargs,
-    ):
-        super().__init__(
-            classes="controller",
-            rounded=(f"{toggle} || 'circle'",),
-            **kwargs,
-        )
-
-        # state initialization
-        self.state.import_pending = False
-
-        # extract trigger name
-        download_export = self.ctrl.trigger_name(export_file_download)
-
-        with self:
-            with v3.VCardTitle(
-                classes=(
-                    f"`d-flex pa-1 position-fixed bg-white ${{ {toggle} ? 'controller-content rounded-t border-b-thin':'rounded-circle'}}`",
-                ),
-                style="z-index: 1;",
-            ):
-                v3.VProgressLinear(
-                    v_if=toggle,
-                    indeterminate=("trame__busy",),
-                    bg_color="rgba(0,0,0,0)",
-                    absolute=True,
-                    color="primary",
-                    location="bottom",
-                    height=2,
-                )
-                v3.VProgressCircular(
-                    v_else=True,
-                    bg_color="rgba(0,0,0,0)",
-                    indeterminate=("trame__busy",),
-                    style="position: absolute; top: 0; left: 0; width: 100%; height: 100%;",
-                    color="primary",
-                    width=3,
-                )
-                v3.VBtn(
-                    icon="mdi-close",
-                    v_if=toggle,
-                    click=f"{toggle} = !{toggle}",
-                    flat=True,
-                    size="sm",
-                )
-                v3.VBtn(
-                    icon="mdi-menu",
-                    v_else=True,
-                    click=f"{toggle} = !{toggle}",
-                    flat=True,
-                    size="sm",
-                )
-                if self.server.hot_reload:
-                    v3.VBtn(
-                        v_show=toggle,
-                        icon="mdi-refresh",
-                        flat=True,
-                        size="sm",
-                        click=self.ctrl.on_server_reload,
-                    )
-                v3.VSpacer()
-                html.Div(
-                    "XArray Viewer",
-                    v_show=toggle,
-                    classes="text-h6 px-2",
-                )
-                v3.VSpacer()
-
-                with v3.VMenu(v_if=toggle, density="compact"):
-                    with html.Template(v_slot_activator="{props}"):
-                        v3.VBtn(
-                            v_bind="props",
-                            icon="mdi-file-arrow-left-right-outline",
-                            flat=True,
-                            size="sm",
-                            classes="mx-1",
-                        )
-                    with v3.VList(density="compact"):
-                        if enable_data_selection:
-                            with v3.VListItem(
-                                title="Export state file",
-                                disabled=("can_load",),
-                                click=f"utils.download('xarray-state.json', trigger('{download_export}'), 'text/plain')",
-                            ):
-                                with html.Template(v_slot_prepend=True):
-                                    v3.VIcon(
-                                        "mdi-cloud-download-outline", classes="mr-n5"
-                                    )
-
-                        with v3.VListItem(
-                            title="Import state file",
-                            click="trame.utils.get('document').querySelector('#fileImport').click()",
-                        ):
-                            html.Input(
-                                id="fileImport",
-                                hidden=True,
-                                type="file",
-                                change=(
-                                    import_file_upload,
-                                    "[$event.target.files]",
-                                ),
-                                __events=["change"],
-                            )
-                            with html.Template(v_slot_prepend=True):
-                                v3.VIcon("mdi-cloud-upload-outline", classes="mr-n5")
-                        v3.VDivider()
-                        with v3.VListItem(
-                            title="Save dataset to disk",
-                            disabled=("can_load",),
-                            click="show_save_dialog = true",
-                        ):
-                            with html.Template(v_slot_prepend=True):
-                                v3.VIcon("mdi-file-download-outline", classes="mr-n5")
-
-            with v3.VCardText(
-                v_show=(toggle, True),
-                classes="controller-content py-1 mt-10",
-            ):
-                if enable_data_selection:
-                    DataOrigin(load_dataset)
-                self.ctrl[xr_update_info] = DataInformation().update_information
-                self.ctrl[source_update_rendering] = RenderingSettings(
-                    source,
-                    update_rendering,
-                ).update_from_source
